requirements.txt
# RAG Agent — pinned to versions in lc-academy-env (verified 2026-03-05)
langgraph==1.0.10
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.17
langchain-chroma==1.1.0
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
pypdf==6.7.4
pydantic==2.12.5
python-dotenv==1.2.1

# Optional: Tavily web-search fallback
# langchain-tavily

# Optional: FlashrankRerank cross-encoder reranking
# flashrank

# Development / testing
pytest>=8.0
